Move filters, pickers and uploaders implementations in a subpackage
