+++ photospicker/picker/pickers/smart_picker.py
